image_classification/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
transformers>=4.42.0
torch>=2.3.0
pillow>=10.3.0
scikit-learn>=1.5.0
tqdm>=4.66.0
python-multipart
huggingface_hub[hf_xet]